requirements.txt
# FastAPI and Server
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.14.0
asyncpg==0.29.0

# Authentication and Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4

# Email
fastapi-mail==1.4.1
jinja2==3.1.3
aiofiles==23.2.1

# Validation and Settings
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Type Checking
mypy==1.14.1
types-python-jose==3.3.4.20240106
types-passlib==1.7.7.20240106
sqlalchemy[mypy]==2.0.25

# Logging
python-json-logger==2.0.7

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Development Tools
black==24.1.0
isort==5.13.2
flake8==7.0.0